raspberry-pi/requirements.txt
pyserial==3.5
google-cloud-firestore==2.11.1